Errors fixed

+++ 1º/Programacion/1er_trimestre/entrega_01/practica_04/EJERCICIOS/src/ejercicio_01.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class ejercicio_01 {
     public static void main(String[] args) {
         
-        final short num = 345;
-        final short per = 20;
-        final short res = num - ((num * per) / 100);
+        final double num = 345;
+        final double per = 20;
+        final double res = num - ((num * per) / 100);
 
         System.out.println(res);
     }
